requirements.txt
certifi==2018.10.15
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch==0.4.1
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.2.1
wincertstore==0.2
